refactor(ui): address PR review — DS consistency, 3-state device status, debounced icon lookup
Follow-up to 89b9fa5 covering all findings from the toolkit review:

DesignSystem invariants:
- Delete dead `Color.dsAccentBadge(forLayoutIndex:)` and centralise the
  EN/RU mapping behind `DSCapsuleTone.layout(forIndex: UInt8)`; narrow
  `IndexBadge.index` to `UInt8` so the firmware wire type carries
  through to the UI boundary.
- Route warning hues through `Color.dsWarn` instead of system `.orange`
  in `DSCard`, `DSErrorBanner`, and the HID Inspector's UNK badge.
- Replace per-capsule `horizontalPadding`/`verticalPadding` knobs with
  a `DSCapsuleSize` enum (`.compact` / `.comfortable` / `.roomy`).
- Add `DSCardTone.accentTinted(Color)` so the HID Inspector's data_type
  badge can stop hand-rolling a `RoundedRectangle` and route through
  `dsCard(...)` like the rest of the app.

Behavioural fixes:
- `SettingsDevicesTab.DeviceRow.liveStatus` distinguishes
  `.invalidProductId` / `.pending` / `.known`, so a typo in
  `config.json` renders a `dsBad` dot with an actionable tooltip
  instead of an indistinguishable grey "Unknown status".
- `AppModel.buildAndStartLinks` now logs and surfaces each filtered-out
  device through `lastSettingsError`, instead of swallowing partial
  failures and only logging when *all* devices failed to parse.
- `HIDInspectorView.categorise` flips its default from `("?", .secondary)`
  to `("UNK", .dsWarn)` so unrecognised data_type bytes (firmware
  reserves `0xB1+`) stand out instead of muting themselves.
- `SettingsAppRulesTab` caches the LaunchServices icon lookup in
  `@State` and refreshes only on commit / row-attach, so typing in
  the bundle-id TextField no longer hits LaunchServices per keystroke.
  Added `.help()` tooltip on the `app.dashed` fallback.

Documentation accuracy:
- Fix `DSCapsuleTone.layoutEN/.layoutRU` doc comments (they render the
  tinted-wash IndexBadge look, not the menubar pill's solid-fill +
  white-glyph look).
- Drop the "matches menubar pill rhythm 7×2" claim — the menubar pill
  is actually 5×1 by design and intentionally does not route through
  `.dsCapsule`.
- Update CLAUDE.md design-system table to reflect the real `.dsCard(...)`
  modifier signature, `IndexBadge(index: UInt8)`, and the new
  `.dsCapsule(tone:size:)` shape.
- Soften the LAYOUT-tag "agree on EN/RU hue" claim in
  `MenuLabel.baseColor` — the inspector tag is always EN-tinted.

+++ RuEnSync/SettingsLayoutsTab.swift
@@ -96,7 +96,7 @@ private struct LayoutRow: View {
                     .foregroundStyle(.tertiary)
                     .frame(width: 14)
 
-                IndexBadge(index: index)
+                IndexBadge(index: UInt8(index))
 
                 VStack(alignment: .leading, spacing: 1) {
                     Text(suffix)
